[Fix] Do not use 'Link All' to ensure that reflection works.
As stated in bug 37242 the sample does not work because the different
properties are read only. This happens when the linker is set to link
everything because the different getter methods are not used.

In this case, we want to keep those methods that are not used so that
the json library, whcih uses reflection, can find readable properties.

To achieve that, we set the linker option to the default value and add
the PreserveAttribute to the properties to keeo them around and, what is
more important, show the users how to fix this kind of situattions if
they want to use the Link All option.