[MemoryDemo] Loading image every time system requests collection cell: * We're doing it purposefully to demonstrate how to *not* deal with memory
Discussion:

UIImage.FromFile method looks in the system caches for an image object with the specified name and returns the variant of that image that is best suited for the main screen. If a matching image object is not already in the cache, this method locates and loads the image data from disk.

On the other hand UIImage .ctor loads image from disk bypassing system cache.