Update TiledPdfView.cs
The app crashes due to a :

`Got a SIGSEGV while executing native code. This usually indicates
a fatal error in the mono runtime or one of the native libraries 
used by your application.` 

After extended debugging an bug was found that was causing the garbage collector to cleanup the back store before we finish using it. This quick fixed the issue.