[SoupChef] Fix build with Xcode 12.2
Xcode 12.2
- adds `arm64` to the simulator architectures
- removes `i386` from the simulator architectures

`lipo` cannot create a _classic_ fat archive with `arm64` simulator and
`arm64` device slices - since they are the same architecture.

This requires `.xcframework` support which is not yet supported by XI
and XM.

However we can tweak the build to do:
- x86_64 only for simulator, which is the only support arch for XI today
- arm64 for devices

and that makes `lipo` happy and we can continue to have this sample under
test automation.

Note: the original `.a` names were incorrect as they did not match the
architectures that were present in the files. I changed them to mention
the platform.